Lots of bug fixes for selecting/changing audio devices, replaced all uses of "inp" with "input" for consistency and clarity, fixes to granstream based on port to webaudio, but granstream does not work yet, new development in apps/sep (in progress).

+++ arco/src/audioio.h
@@ -40,7 +40,7 @@ extern int aud_state;
 extern int aud_quit_request;
 
 extern int64_t arco_frames_done;
-extern int64_t arco_blocks_done;
+extern int arco_blocks_done;
 
 void audioio_initialize();
 void audioio_finish(Bridge_info *bridge);
